requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SQLAlchemy==3.1.1
python-dateutil==2.8.2
reportlab==4.0.7
Werkzeug==3.0.1
openpyxl==3.1.2
pandas==2.1.4



